GSYE-333: Moved EnergyParameters classes to their own Python package,
