+++ src/gsy_e/models/strategy/predefined_load.py
@@ -18,62 +18,10 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 from typing import Union
 
 from gsy_framework.constants_limits import ConstSettings
-from gsy_framework.utils import find_object_of_same_weekday_and_time
-from pendulum import DateTime, duration
+from pendulum import duration
 
-from gsy_e.gsy_e_core.exceptions import GSyException
-from gsy_e.models.state import LoadState
-from gsy_e.models.strategy.load_hours import LoadHoursPerDayEnergyParameters, LoadHoursStrategy
-from gsy_e.models.strategy.profile import EnergyProfile
-
-
-class DefinedLoadEnergyParameters(LoadHoursPerDayEnergyParameters):
-    """Energy parameters for the defined load strategy class."""
-    def __init__(self, daily_load_profile=None, daily_load_profile_uuid: str = None):
-        super().__init__(avg_power_W=0, hrs_per_day=24, hrs_of_day=list(range(0, 24)))
-        self._energy_profile = EnergyProfile(daily_load_profile, daily_load_profile_uuid)
-        self.state = LoadState()
-
-    def serialize(self):
-        return {
-            "daily_load_profile": self._energy_profile.input_profile,
-            "daily_load_profile_uuid": self._energy_profile.input_profile_uuid
-        }
-
-    def event_activate_energy(self, area):
-        """
-        Runs on activate event.
-        :return: None
-        """
-        self._energy_profile.read_or_rotate_profiles()
-        super().event_activate_energy(area)
-
-    def reset(self, time_slot: DateTime, **kwargs) -> None:
-        if kwargs.get("daily_load_profile") is not None:
-            self._energy_profile.input_profile = kwargs["daily_load_profile"]
-            self._energy_profile.read_or_rotate_profiles(reconfigure=True)
-
-    def update_energy_requirement(self, time_slot, overwrite=False):
-        if not self._energy_profile.profile:
-            raise GSyException(
-                "Load tries to set its energy forecasted requirement "
-                "without a profile.")
-        load_energy_kwh = find_object_of_same_weekday_and_time(
-            self._energy_profile.profile, time_slot)
-        self.state.set_desired_energy(load_energy_kwh * 1000, time_slot, overwrite=False)
-        self.state.update_total_demanded_energy(time_slot)
-
-    def _operating_hours(self, energy_kWh):
-        """
-        Disabled feature for this subclass
-        """
-        return 0
-
-    def allowed_operating_hours(self, time_slot):
-        """
-        Disabled feature for this subclass
-        """
-        return True
+from gsy_e.models.strategy.load_hours import LoadHoursStrategy
+from gsy_e.models.strategy.energy_parameters.load import DefinedLoadEnergyParameters
 
 
 class DefinedLoadStrategy(LoadHoursStrategy):
