+++ src/gsy_e/models/strategy/predefined_pv.py
@@ -15,86 +15,13 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
-import pathlib
 
 from gsy_framework.constants_limits import ConstSettings
-from gsy_framework.read_user_profile import InputProfileTypes, read_arbitrary_profile
-from gsy_framework.utils import (convert_kW_to_kWh, find_object_of_same_weekday_and_time,
-                                 key_in_dict_and_not_none)
 from pendulum import duration
 
-from gsy_e.gsy_e_core.exceptions import GSyException
-from gsy_e.gsy_e_core.util import d3a_path
-from gsy_e.models.strategy.profile import EnergyProfile
-from gsy_e.models.strategy.pv import PVEnergyParameters, PVStrategy
-
-
-class PVPredefinedEnergyParameters(PVEnergyParameters):
-    """Energy-related parameters for the PVPredefinedStrategy class."""
-    def __init__(self, panel_count: int = 1, cloud_coverage: int = None,
-                 capacity_kW: float = None):
-        super().__init__(panel_count, capacity_kW)
-        self.cloud_coverage = cloud_coverage
-        self._power_profile_index = cloud_coverage
-        self.energy_profile = {}
-
-    def serialize(self):
-        return {
-            **super().serialize(),
-            "cloud_coverage": self.cloud_coverage
-        }
-
-    def read_predefined_profile_for_pv(self, simulation_config):
-        """
-        Reads profile data from the predefined power profiles. Reads config and constructor
-        parameters and selects the appropriate predefined profile.
-        """
-        if self._power_profile_index is None or self._power_profile_index == 4:
-            if simulation_config.pv_user_profile is not None:
-                self.energy_profile = simulation_config.pv_user_profile
-                return
-            self._power_profile_index = simulation_config.cloud_coverage
-        if self._power_profile_index == 0:  # 0:sunny
-            profile_path = (
-                pathlib.Path(d3a_path + "/resources/Solar_Curve_sunny_normalized.csv"))
-        elif self._power_profile_index == 1:  # 1:partial
-            profile_path = (
-                pathlib.Path(d3a_path + "/resources/Solar_Curve_partial_normalized.csv"))
-        elif self._power_profile_index == 2:  # 2:cloudy
-            profile_path = (
-                pathlib.Path(d3a_path + "/resources/Solar_Curve_cloudy_normalized.csv"))
-        else:
-            raise ValueError("Energy_profile has to be in [0,1,2,4]")
-
-        power_weight_profile = read_arbitrary_profile(
-            InputProfileTypes.IDENTITY, profile_path)
-
-        self.energy_profile = {
-            time_slot: convert_kW_to_kWh(
-                weight * self.capacity_kW, simulation_config.slot_length)
-            for time_slot, weight in power_weight_profile.items()}
-
-    def set_produced_energy_forecast_in_state(
-            self, config_cloud_coverage, owner_name, time_slots, reconfigure=True):
-        """Update the production energy forecast."""
-        self._power_profile_index = self.cloud_coverage or config_cloud_coverage
-        if not self.energy_profile:
-            raise GSyException(
-                f"PV {owner_name} tries to set its available energy forecast without a "
-                "power profile.")
-        for time_slot in time_slots:
-            available_energy_kWh = find_object_of_same_weekday_and_time(
-                self.energy_profile, time_slot) * self.panel_count
-            self._state.set_available_energy(available_energy_kWh, time_slot, reconfigure)
-
-    def reconfigure(self, **kwargs):
-        """Reconfigure the device properties at runtime using the provided arguments."""
-
-        # kwargs["cloud_coverage"] = None is a valid value, therefore a None check should not
-        # be added here.
-        if "cloud_coverage" in kwargs:
-            self.cloud_coverage = kwargs["cloud_coverage"]
-            self._power_profile_index = self.cloud_coverage
+from gsy_e.models.strategy.energy_parameters.pv import (
+    PVPredefinedEnergyParameters, PVUserProfileEnergyParameters)
+from gsy_e.models.strategy.pv import PVStrategy
 
 
 class PVPredefinedStrategy(PVStrategy):
@@ -171,46 +98,6 @@ class PVPredefinedStrategy(PVStrategy):
         self._energy_params.reconfigure(**kwargs)
         self._energy_params.read_predefined_profile_for_pv(self.simulation_config)
         super().area_reconfigure_event(**kwargs)
-
-
-class PVUserProfileEnergyParameters(PVEnergyParameters):
-    """Energy-related parameters for the PVUserProfile Strategy class."""
-    def __init__(self, panel_count: int = 1, power_profile: str = None,
-                 power_profile_uuid: str = None):
-        super().__init__(panel_count, None)
-        self._energy_profile = EnergyProfile(power_profile, power_profile_uuid)
-
-    def serialize(self):
-        return {
-            **super().serialize(),
-            "power_profile": self._energy_profile.input_profile,
-            "power_profile_uuid": self._energy_profile.input_profile_uuid
-        }
-
-    def read_predefined_profile_for_pv(self):
-        """
-        Reads profile data from the power profile. Handles csv files and dicts.
-        :return: key value pairs of time to energy in kWh
-        """
-        self._energy_profile.read_or_rotate_profiles()
-
-    def reset(self, **kwargs):
-        """Reset the energy parameters of the strategy."""
-        if key_in_dict_and_not_none(kwargs, "power_profile"):
-            self._energy_profile.input_profile = kwargs["power_profile"]
-        self._energy_profile.read_or_rotate_profiles(reconfigure=True)
-
-    def set_produced_energy_forecast_in_state(
-            self, owner_name, time_slots, reconfigure=True):
-        """Update the production energy forecast."""
-        if not self._energy_profile.profile:
-            raise GSyException(
-                f"PV {owner_name} tries to set its available energy forecast without a "
-                "power profile.")
-        for time_slot in time_slots:
-            available_energy_kWh = find_object_of_same_weekday_and_time(
-                self._energy_profile.profile, time_slot) * self.panel_count
-            self._state.set_available_energy(available_energy_kWh, time_slot, reconfigure)
 
 
 class PVUserProfileStrategy(PVStrategy):
