+++ src/gsy_e/models/strategy/pv.py
@@ -15,97 +15,29 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
-import math
 import traceback
 from logging import getLogger
 
-import pendulum
 from gsy_framework.constants_limits import ConstSettings
 from gsy_framework.exceptions import GSyException
 from gsy_framework.read_user_profile import read_arbitrary_profile, InputProfileTypes
 from gsy_framework.utils import (
-    convert_kW_to_kWh, find_object_of_same_weekday_and_time, key_in_dict_and_not_none)
+    find_object_of_same_weekday_and_time, key_in_dict_and_not_none)
 from gsy_framework.validators import PVValidator
 from pendulum import duration  # noqa
-from pendulum.datetime import DateTime
 
 from gsy_e import constants
 from gsy_e.gsy_e_core.exceptions import MarketException
 from gsy_e.gsy_e_core.util import get_market_maker_rate_from_config
-from gsy_e.models.state import PVState
 from gsy_e.models.base import AssetType
-from gsy_e.models.strategy import BidEnabledStrategy, utils
+from gsy_e.models.state import PVState
+from gsy_e.models.strategy import BidEnabledStrategy
+from gsy_e.models.strategy.energy_parameters.pv import PVEnergyParameters
 from gsy_e.models.strategy.future.strategy import future_market_strategy_factory
 from gsy_e.models.strategy.settlement.strategy import settlement_market_strategy_factory
 from gsy_e.models.strategy.update_frequency import TemplateStrategyOfferUpdater
 
 log = getLogger(__name__)
-
-
-class PVEnergyParameters:
-    """Energy parameters for the PV strategy with gaussian generation profile."""
-    def __init__(self, panel_count: int = 1, capacity_kW: float = None):
-        PVValidator.validate_energy(panel_count=panel_count, capacity_kW=capacity_kW)
-
-        self.panel_count = panel_count
-        self.capacity_kW = capacity_kW
-        self._state = PVState()
-
-    def serialize(self):
-        """Return dict with the current energy parameter values."""
-        return {
-            "panel_count": self.panel_count,
-            "capacity_kW": self.capacity_kW
-        }
-
-    def activate(self, simulation_config):
-        """Activate the energy parameters, called during the event_activate strategy event."""
-        if self.capacity_kW is None:
-            self.capacity_kW = simulation_config.capacity_kW
-
-    def set_produced_energy_forecast(self, time_slot, slot_length, reconfigure=True):
-        """Generate the energy forecast value for the specified timeslot."""
-        difference_to_midnight_in_minutes = time_slot.diff(
-            pendulum.datetime(
-                time_slot.year, time_slot.month, time_slot.day)).in_minutes() % (60 * 24)
-        available_energy_kWh = self._gaussian_energy_forecast_kWh(
-            slot_length, difference_to_midnight_in_minutes) * self.panel_count
-        self._state.set_available_energy(available_energy_kWh, time_slot, reconfigure)
-
-    def reset(self, **kwargs):
-        """Reset / update energy parameters."""
-        if key_in_dict_and_not_none(kwargs, "panel_count"):
-            self.panel_count = kwargs["panel_count"]
-        if key_in_dict_and_not_none(kwargs, "capacity_kW"):
-            self.capacity_kW = kwargs["capacity_kW"]
-
-    def _gaussian_energy_forecast_kWh(self, slot_length, time_in_minutes=0):
-        # The sun rises at approx 6:30 and sets at 18hr
-        # time_in_minutes is the difference in time to midnight
-
-        # Clamp to day range
-        time_in_minutes %= 60 * 24
-
-        if (8 * 60) > time_in_minutes or time_in_minutes > (16.5 * 60):
-            gauss_forecast = 0
-
-        else:
-            gauss_forecast = self.capacity_kW * math.exp(
-                # time/5 is needed because we only have one data set per 5 minutes
-
-                (- (((round(time_in_minutes / 5, 0)) - 147.2)
-                    / 38.60) ** 2
-                 )
-            )
-
-        return round(convert_kW_to_kWh(gauss_forecast, slot_length), 4)
-
-    def set_energy_measurement_kWh(self, time_slot: DateTime) -> None:
-        """Set the (simulated) actual energy produced by the device in a market slot."""
-        energy_forecast_kWh = self._state.get_energy_production_forecast_kWh(time_slot)
-        simulated_measured_energy_kWh = utils.compute_altered_energy(energy_forecast_kWh)
-
-        self._state.set_energy_measurement_kWh(simulated_measured_energy_kWh, time_slot)
 
 
 class PVStrategy(BidEnabledStrategy):
